+++ src/linter/VerilatorLinter.ts
@@ -99,7 +99,9 @@ export default class VerilatorLinter extends BaseLinter {
     let command: string = binPath + ' ' + args.join(' ');
 
     let cwd: string = this.runAtFileLocation
-      ? docFolder
+      ? isWindows
+        ? path.dirname(docUri)
+        : docFolder
       : vscode.workspace.workspaceFolders[0].uri.fsPath;
 
     this.logger.info('[verilator] Execute');